requirements.txt
huggingface_hub==1.6.0
langchain==1.2.12
langchain-classic==1.0.2
langchain-community==0.4.1
langchain-core==1.2.18
langchain-huggingface==1.2.1
langchain-ollama==1.0.1
langchain-openai==1.1.11
langchain-pinecone==0.2.13
langchain-text-splitters==1.1.1
ollama==0.6.1
pinecone==7.3.0
python-dotenv==1.2.2
sentence-transformers==5.2.3
streamlit==1.55.0
torch==2.10.0
transformers==5.3.0